fix: make layout visualizer mergeable
