+++ docs/internal/processes/factory-replay-relevant-files.md
@@ -69,10 +69,11 @@ that consumes the replay kernel.
   dashboard-specific Factory-world event reducer; `projectSnapshot.ts` owns
   its projection. Keep browser state, checkpoints, persistence, and Zustand
   outside `packages/factory-replay`.
-- `ui/packages/factory-visualizers/src/factory-topology-replay.tsx` converts
-  caller-validated `FactoryVisualizationLayoutV1` annotations into separate,
-  inert React Flow nodes. Keep them out of canonical topology layout and edge
-  routing; visibility may remove only those projected nodes so React Flow
+- `ui/packages/factory-visualizers/src/factory-topology-flow-projection.ts`
+  derives caller-validated `FactoryVisualizationLayoutV1` annotations as
+  separate, inert React Flow nodes; `factory-topology-replay-nodes.tsx` owns
+  their read-only node and image rendering. Keep them out of canonical topology
+  layout and edge routing; visibility may remove only those projected nodes so React Flow
   fit-to-view remains derived from what is currently visible. Embedded raster
   annotations must decode into Blob URLs, revoke those URLs on replacement,
   removal, image failure, and unmount, and contain preparation failures in the
